Add search event to keyup handling block
According to [this StackOverflow answer](http://stackoverflow.com/questions/2977023/how-do-you-detect-the-clearing-of-a-search-html5-input), clicking on the ⓧ  issues a `search` event from the input.

Just added the `search` event to the `keyup` handling for the event. Confirmed by debugging the block that this isn't double called. `search` is only triggered when the ⓧ  is clicked, not on every keypress. `keyup` works normally.